+++ src/frontend/components/Dashboard/DashboardVideoWrapper/index.spec.tsx
@@ -174,18 +174,16 @@ describe('<DashboardVideoWrapper />', () => {
           room_name: 'jitsi_conference',
         },
       },
-      upload_state: uploadState.HARVESTING,
+      upload_state: uploadState.READY,
     });
 
     render(
       wrapInIntlProvider(wrapInRouter(<DashboardVideoWrapper video={video} />)),
     );
 
-    screen.getByRole('link', { name: 'Dashboard' });
+    await screen.findByRole('link', { name: 'Dashboard' });
     screen.getByRole('link', { name: 'Playlist' });
 
-    screen.getByText(
-      'Your video is currently converting from a live video to a VOD. This may take up to an hour. You can close the window and come back later.',
-    );
+    screen.getByText('Your video is ready to play.');
   });
 });