+++ src/frontend/components/TeacherLiveLifecycleControls/StopLiveButton/index.tsx
@@ -1,10 +1,13 @@
-import { Box, Button, ButtonProps } from 'grommet';
+import { Box, Button, ButtonProps, Heading, Paragraph, Spinner } from 'grommet';
 import { normalizeColor } from 'grommet/utils';
-import React from 'react';
+import React, { Fragment, useEffect, useState } from 'react';
+import toast from 'react-hot-toast';
 import { defineMessages, useIntl } from 'react-intl';
 
 import { StopSVG } from 'components/SVGIcons/StopSVG';
-import { useStopLiveConfirmation } from 'data/stores/useStopLiveConfirmation';
+import { stopLive } from 'data/sideEffects/stopLive';
+import { useLiveModaleConfiguration } from 'data/stores/useLiveModale';
+import { useVideo } from 'data/stores/useVideo';
 import { Video } from 'types/tracks';
 import { theme } from 'utils/theme/theme';
 
@@ -14,7 +17,40 @@ const messages = defineMessages({
     description: 'Title for the button to end streaming as a teacher.',
     id: 'components.DashboardVideoLiveEndButton.endLive',
   },
+  noTitle: {
+    defaultMessage: 'No title',
+    description: 'Title placeholder when no title is defined for this live',
+    id: 'components.DashboardVideoLiveEndButton.noTitle',
+  },
+  alertInfo: {
+    defaultMessage:
+      'You are about to stop the live.{br}Every viewers will be disconnected.{br}Are you sure you want to stop this Live ?{br}Beware, live is still running until stopped.',
+    description:
+      'Message shows when the teacher needs to confirm he wants to stop the live.',
+    id: 'components.DashboardVideoLiveEndButton.alertInfo',
+  },
+  stopButtonTitle: {
+    defaultMessage: 'Stop the live',
+    description: 'Stop Button title',
+    id: 'components.DashboardVideoLiveEndButton.stopButtonTitle',
+  },
+  error: {
+    defaultMessage: 'An error occured, please try again.',
+    description:
+      'Error message displayed when and error occured when trying to pause',
+    id: 'components.DashboardVideoLiveEndButton.error',
+  },
+  cancel: {
+    defaultMessage: 'Cancel',
+    description: 'Button label to cancel stopping a live',
+    id: 'components.DashboardVideoLiveEndButton.cancel',
+  },
 });
+
+type EndLiveStatus =
+  | { type: 'idle' }
+  | { type: 'loading' }
+  | { type: 'error'; error: unknown };
 
 interface StopLiveButtonProps extends ButtonProps {
   video: Video;
@@ -22,26 +58,111 @@ interface StopLiveButtonProps extends ButtonProps {
 
 export const StopLiveButton = ({ video, ...props }: StopLiveButtonProps) => {
   const intl = useIntl();
-  const [shouldShowStopAler, setShouldShowStopAlert] =
-    useStopLiveConfirmation();
+  const [liveModaleConfiguration, setLiveModaleConfiguration] =
+    useLiveModaleConfiguration();
+  const [status, setStatus] = useState<EndLiveStatus>({ type: 'idle' });
+  const { updateVideo } = useVideo((state) => ({
+    updateVideo: state.addResource,
+  }));
+
+  useEffect(() => {
+    if (status.type === 'error') {
+      toast.error(intl.formatMessage(messages.error));
+      setStatus({ type: 'idle' });
+    }
+  }, [status, intl, messages]);
+
+  useEffect(() => {
+    if (status.type !== 'loading') {
+      return;
+    }
+
+    let cancel = false;
+    const apiCall = async () => {
+      try {
+        const updatedVideo = await stopLive(video);
+        if (cancel) {
+          return;
+        }
+        setStatus({ type: 'idle' });
+        updateVideo(updatedVideo);
+      } catch (error) {
+        setStatus({ type: 'error', error });
+      }
+    };
+
+    apiCall();
+    return () => {
+      cancel = true;
+    };
+  }, [status, updateVideo]);
+
+  const title = video.title || intl.formatMessage(messages.noTitle);
+  const modaleContent = (
+    <Fragment>
+      <Heading
+        a11yTitle={title}
+        color="white"
+        level="3"
+        margin={{ horizontal: 'auto' }}
+        title={title}
+        truncate
+        textAlign="center"
+        style={{ width: '100%' }}
+      >
+        {title}
+      </Heading>
+      <Paragraph color="white" textAlign="center">
+        {intl.formatMessage(messages.alertInfo, { br: <br /> })}
+      </Paragraph>
+    </Fragment>
+  );
 
   return (
     <Button
       primary
       color={normalizeColor('red-active', theme)}
-      disabled={shouldShowStopAler}
+      disabled={!!liveModaleConfiguration || status.type === 'loading'}
       label={
         <Box flex direction="row" style={{ whiteSpace: 'nowrap' }}>
           {intl.formatMessage(messages.title)}
-          <StopSVG
-            iconColor="white"
-            height="25px"
-            width="25px"
-            containerStyle={{ margin: 'auto', marginLeft: '8px' }}
-          />
+          {status.type !== 'loading' && (
+            <StopSVG
+              iconColor="white"
+              height="25px"
+              width="25px"
+              containerStyle={{ margin: 'auto', marginLeft: '8px' }}
+            />
+          )}
+          {status.type === 'loading' && (
+            <Spinner
+              data-testid="loader-id"
+              color="white"
+              margin={{ left: 'small' }}
+            />
+          )}
         </Box>
       }
-      onClick={() => setShouldShowStopAlert(true)}
+      onClick={() =>
+        setLiveModaleConfiguration({
+          content: modaleContent,
+          actions: [
+            {
+              label: intl.formatMessage(messages.stopButtonTitle),
+              action: () => {
+                setStatus({ type: 'loading' });
+                setLiveModaleConfiguration(null);
+              },
+            },
+            {
+              label: intl.formatMessage(messages.cancel),
+              action: () => {
+                setLiveModaleConfiguration(null);
+              },
+            },
+          ],
+        })
+      }
       {...props}
     />
   );