+++ src/frontend/components/VideoPlayer/index.tsx
@@ -1,25 +1,15 @@
 import { Stack } from 'grommet';
-import React, { useRef, useState, useEffect } from 'react';
+import React, { useRef, useEffect } from 'react';
 import { Redirect } from 'react-router';
 
 import { FULL_SCREEN_ERROR_ROUTE } from 'components/ErrorComponents/route';
-import { PublicPausedLiveVideo } from 'components/PublicPausedLiveVideo';
 
 import { useThumbnail } from 'data/stores/useThumbnail';
 import { useTimedTextTrackLanguageChoices } from 'data/stores/useTimedTextTrackLanguageChoices';
 import { useVideoProgress } from 'data/stores/useVideoProgress';
 import { createPlayer } from 'Player/createPlayer';
-import {
-  liveState,
-  TimedText,
-  timedTextMode,
-  Video,
-  videoSize,
-} from 'types/tracks';
-import { VideoPlayerInterface } from 'types/VideoPlayer';
-import { resumeLive } from 'utils/resumeLive';
+import { TimedText, timedTextMode, Video, videoSize } from 'types/tracks';
 import { Nullable } from 'utils/types';
-import { useAsyncEffect } from 'utils/useAsyncEffect';
 import { useIntl } from 'react-intl';
 
 const trackTextKind: { [key in timedTextMode]?: string } = {
@@ -41,12 +31,7 @@ const VideoPlayer = ({
   defaultVolume,
 }: BaseVideoPlayerProps) => {
   const intl = useIntl();
-  const [player, setPlayer] = useState<VideoPlayerInterface>();
-  const videoNodeRef = useRef(null as Nullable<HTMLVideoElement>);
-  const [isLivePausedOrStopping, setIsLivePausedOrStopping] = useState(
-    video?.live_state &&
-      [liveState.STOPPING, liveState.PAUSED].includes(video.live_state),
-  );
+  const videoNodeRef = useRef<Nullable<HTMLVideoElement>>(null);
 
   const { choices, getChoices } = useTimedTextTrackLanguageChoices(
     (state) => state,
@@ -90,7 +75,6 @@ const VideoPlayer = ({
           }
         },
       );
-      setPlayer(createdPlayer);
 
       document.dispatchEvent(
         new CustomEvent('marsha_player_created', {
@@ -104,27 +88,6 @@ const VideoPlayer = ({
       return () => createdPlayer && createdPlayer.destroy();
     }
   }, []);
-
-  useAsyncEffect(async () => {
-    // live is resuming
-    if (video?.live_state === liveState.STARTING && player) {
-      setIsLivePausedOrStopping(true);
-    }
-
-    if (
-      video?.live_state &&
-      [liveState.STOPPING, liveState.PAUSED].includes(video.live_state)
-    ) {
-      setIsLivePausedOrStopping(true);
-    }
-
-    // live is running and the player is ready, we check if the hls manifest contains frame to display
-    if (video?.live_state === liveState.RUNNING && player) {
-      await resumeLive(video);
-      player.setSource(player.getSource());
-      setIsLivePausedOrStopping(false);
-    }
-  }, [video?.live_state, player, setIsLivePausedOrStopping]);
 
   // The video is somehow missing and jwt must be set
   if (!video) {
@@ -177,12 +140,6 @@ const VideoPlayer = ({
             />
           ))}
       </video>
-      {isLivePausedOrStopping && player && (
-        <PublicPausedLiveVideo
-          video={video}
-          videoNodeRef={videoNodeRef.current!}
-        />
-      )}
     </Stack>
   );
 };